+++ src/app/(app)/transfers/actions.ts
@@ -2,30 +2,49 @@
 
 import { revalidatePath } from 'next/cache'
 
+import { auth } from '@/auth'
 import prisma from '@/lib/db'
 import { createTransferSchema, updateTransferSchema } from '@/schemas/transfer.schema'
 
-// Currently not by userId
 export async function getUserTransfers() {
+   const session = await auth()
+
+   if (!session?.user.id) {
+      return { message: 'Not authenticated!', status: 403 }
+   }
+
    const userTransfers = await prisma.transfer.findMany({
+      where: { userId: { equals: session.user.id } },
       include: {
          accountFrom: { select: { name: true, id: true } },
          accountTo: { select: { name: true, id: true } },
       },
    })
 
-   return { transfers: userTransfers }
+   return { transfers: userTransfers, status: 200 }
 }
 
 export async function getOneTransfer(id: string) {
+   const session = await auth()
+
+   if (!session?.user.id) {
+      return { message: 'Not authenticated!', status: 403 }
+   }
+
    const transfer = await prisma.transfer.findUnique({
       where: { id },
    })
 
-   return { transfer }
+   return { transfer, status: 200 }
 }
 
 export async function addTransfer(formData: FormData) {
+   const session = await auth()
+
+   if (!session?.user.id) {
+      return { message: 'Not authenticated!', status: 403 }
+   }
+
    const data = {
       title: formData.get('title'),
       amount: formData.get('amount'),
@@ -38,15 +57,22 @@ export async function addTransfer(formData: FormData) {
    const validate = createTransferSchema.parse(data)
 
    const newTransfer = await prisma.transfer.create({
-      data: validate,
+      data: { ...validate, userId: session.user.id },
    })
 
    revalidatePath('/transfers')
    revalidatePath('/dashboard')
-   return { message: 'Successfully created!', transfer: newTransfer }
+
+   return { message: 'Successfully created!', status: 200, transfer: newTransfer }
 }
 
 export async function updateTransfer(formData: FormData) {
+   const session = await auth()
+
+   if (!session?.user.id) {
+      return { message: 'Not authenticated!', status: 403 }
+   }
+
    const data = {
       id: formData.get('id'),
       title: formData.get('title'),
@@ -66,15 +92,23 @@ export async function updateTransfer(formData: FormData) {
 
    revalidatePath('/transfers')
    revalidatePath('/dashboard')
-   return { message: 'Successfully updated!', transfer: updatedTransfer }
+
+   return { message: 'Successfully updated!', staus: 200, transfer: updatedTransfer }
 }
 
 export async function deleteTransfer(id: string) {
+   const session = await auth()
+
+   if (!session?.user.id) {
+      return { message: 'Not authenticated!', status: 403 }
+   }
+
    const deletedTransfer = await prisma.transfer.delete({
       where: { id },
    })
 
    revalidatePath('/transfers')
    revalidatePath('/dashboard')
-   return { message: 'Transfer deleted', transfer: deletedTransfer }
+
+   return { message: 'Transfer deleted', status: 200, transfer: deletedTransfer }
 }
